We have added the real length (minus the status bits) to the status bits.
