+++ hat/hat-core/src/main/java/hat/buffer/Buffer.java
@@ -26,8 +26,8 @@ package hat.buffer;
 
 
 import hat.ifacemapper.BoundSchema;
+import hat.ifacemapper.BufferState;
 import hat.ifacemapper.MappableIface;
-import hat.ifacemapper.SegmentMapper;
 
 import java.lang.foreign.MemoryLayout;
 import java.lang.foreign.MemorySegment;
@@ -41,21 +41,21 @@ import static hat.ifacemapper.MapperUtil.SECRET_SEGMENT_METHOD_NAME;
 public interface Buffer extends MappableIface {
 
     default boolean isDeviceDirty(){
-        return SegmentMapper.BufferState.of(this).isDeviceDirty();
+        return BufferState.of(this).isDeviceDirty();
     }
     default boolean isHostChecked(){
-        return SegmentMapper.BufferState.of(this).isHostChecked();
+        return BufferState.of(this).isHostChecked();
     }
 
     default void clearDeviceDirty(){
-         SegmentMapper.BufferState.of(this).clearDeviceDirty();
+         BufferState.of(this).clearDeviceDirty();
     }
     default void setHostDirty(){
-        SegmentMapper.BufferState.of(this).setHostDirty(true);
+        BufferState.of(this).setHostDirty(true);
     }
 
     default void setHostChecked(){
-        SegmentMapper.BufferState.of(this).setHostChecked(true);
+        BufferState.of(this).setHostChecked(true);
     }
 
     interface Union extends MappableIface {
